requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.40.0
psycopg2-binary>=2.9.0
litellm>=1.51.0
python-dotenv>=1.0.0
